+++ CHANGELOG.md
@@ -37,6 +37,7 @@ The new system has a number of benefits
 * Add flake8 plugins: `flake8-pytest-style` and `flake8-annotations`
 * Add additional tests
 * Improve existing tests
+* Add tests and samples for Ubuntu 12.04
 * Add samples for NetBSD (support coming in future release)